backend/pyproject.toml
[project]
name = "rokr-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy==2.0.37",
    "alembic==1.14.0",
    "pydantic==2.10.5",
    "pydantic-settings==2.7.1",
    "psycopg2-binary==2.9.10",
    "python-jose[cryptography]==3.3.0",
    "httpx==0.28.1",
    "python-multipart==0.0.20",
    "celery==5.4.0",
    "redis==5.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
    "pytest-cov==6.0.0",
    "httpx==0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
